First attempt at backprop
TODO: 1. adjust all functions to use the same data type as the network's parameters. 2. rewrite network using weights and biases as separate arrays, rather than together in a single parameters matrix. The forward and backward propagation get unnecessarily complex because of the need to add columns of 1s any time a bias term is being matrix multiplied with an input matrix. The beauty of the algorithm begins to be lost in trying to manage the array shapes through concatenation and transposition.